+++ package.json
@@ -14,8 +14,7 @@
     "benchmark:erc20": "ts-node scripts/03_benchmark.ts",
     "benchmark:trades_gas": "ts-node scripts/04_benchmark.ts",
     "benchmark:cancel": "ts-node scripts/05_benchmark.ts",
-    "benchmark:cancel_gas": "ts-node scripts/06_benchmark.ts",
-    "benchmark:batch_cancel_gas": "ts-node scripts/07_benchmark.ts"
+    "benchmark:batch_cancel_gas": "ts-node scripts/06_benchmark.ts"
   },
   "repository": {
     "type": "git",
